+++ Source/Kernel/Grains.Interfaces/Jobs/FailedResumingJobSteps.cs
@@ -1,0 +1,11 @@
+// Copyright (c) Cratis. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Cratis.Chronicle.Concepts.Jobs;
+namespace Cratis.Chronicle.Grains.Jobs;
+
+/// <summary>
+/// The job steps that failed resuming.
+/// </summary>
+/// <param name="FailedJobSteps">The job step ids.</param>
+public record FailedResumingJobSteps(IEnumerable<JobStepId> FailedJobSteps);
